# packages/api-spec/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Manhaj API",
    "version": "1.0.0",
    "description": "Offline-First Student Question Bank API"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "authToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-auth-token"
      }
    },
...
  },
  "paths": {
...
    "/auth/refresh": {
      "post": {
        "summary": "Refresh access token",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully"
          }
        }
      }
    },
...
  },
...
}